+++ modules/axfs/src/lib.rs
@@ -1,6 +1,9 @@
 #![no_std]
 
+pub mod devfs;
 pub mod memfs;
+pub mod mount;
+pub mod procfs;
 
 pub use axvfs::{FileType, InodeId, Metadata, VfsError, VfsOps};